Handling any error or exception

+++ server/src/main/java/io/littlehorse/common/model/getable/core/wfrun/ThreadRunModel.java
@@ -510,7 +510,7 @@ public class ThreadRunModel extends LHSerializable<ThreadRun> {
         thr.handlingFailure = new HandlingFailureHaltReasonModel();
         thr.handlingFailure.handlerThreadId = handlerThreadNumber;
 
-        childThreadIds.add((Integer) handlerThreadNumber);
+        childThreadIds.add(handlerThreadNumber);
         haltReasons.add(thr);
     }
